create-profile

+++ src/party/user/user.orchestrator.ts
@@ -17,6 +17,8 @@ import { createAccessRoleRepositoryFactory } from "../../authorization/access-ro
 import { AccessRole } from "../../authorization/access-role/access.role";
 import { CredentialRepository} from "../../authentication/credential/credential.repository";
 import { createCredentialRepositoryFactory } from "../../authentication/credential/credential.repository.factory";
+import { SessionRepository } from "../../session/session.repository";
+import { createSessionRepositoryFactory } from "../../session/session.repository.factory";
 
 export class UserOrchestrator {
 
@@ -24,12 +26,14 @@ export class UserOrchestrator {
   private credentialRepository: CredentialRepository;
   private partyAccessRoleRepository: PartyAccessRoleRepository;
   private accessRoleRepository: AccessRoleRepository;
+  private sessionRepository: SessionRepository;
 
   constructor() {
     this.userRepository = createUserRepository();
     this.credentialRepository = createCredentialRepositoryFactory();
     this.partyAccessRoleRepository = createPartyAccessRoleRepositoryFactory();
     this.accessRoleRepository = createAccessRoleRepositoryFactory();
+    this.sessionRepository = createSessionRepositoryFactory();
   }
 
 
@@ -50,7 +54,23 @@ export class UserOrchestrator {
         }));
     }
 
-  getUser (partyId: string): Observable<UserResponse> {
+  getUser (partyId: string, sessionId?: string): Observable<UserResponse> {
+      if (partyId === "me") {
+          return this.sessionRepository.getSessionById(sessionId)
+              .pipe(switchMap(session => {
+                  if (!session) {
+                      return of(undefined);
+                  } else {
+                      return this.getUserLocal(session.partyId);
+                  }
+              }));
+      } else {
+          return this.getUserLocal(partyId);
+      }
+  }
+
+  getUserLocal(partyId: string): Observable<UserResponse> {
+      console.log(partyId);
       return this.userRepository.getUser(partyId)
           .pipe(switchMap(user => {
               if (!user) return of(undefined);
@@ -61,7 +81,7 @@ export class UserOrchestrator {
                       if (accessRoleIds.length < 1)  return of(new UserResponse(user, partyAccessRoles));
                       return this.accessRoleRepository.getAccessRoleByIds(accessRoleIds)
                           .pipe(map( accessRoles => {
-                             if (accessRoles.length < 1) return new UserResponse(user, partyAccessRoles);
+                              if (accessRoles.length < 1) return new UserResponse(user, partyAccessRoles);
                               partyAccessRoles.forEach( value => {
                                   const index = accessRoles.findIndex(x => x.accessRoleId === value.accessRoleId);
                                   value.accessRole = index !== -1 ? accessRoles[index] : new AccessRole();
@@ -70,33 +90,39 @@ export class UserOrchestrator {
                           }));
                   }));
           }));
-    }
+  }
 
-  saveUser (user: User, partyAccessRoles: PartyAccessRole[]): Observable<User> {
-      const credential = JSON.parse(JSON.stringify(new Credential().toJson()));
-      credential.username = user.username;
-      credential.password = generate({
-          length: 10,
-          numbers: true
-      });
-      return this.credentialRepository.addCredential(credential)
-          .pipe(switchMap( credentialRes => {
-             if (!credentialRes) return of(undefined);
-             user.username = undefined;
-             user.partyId = credentialRes.credential.partyId;
-              return this.userRepository.saveUser(user)
-                  .pipe(switchMap( userRes => {
-                      if (!userRes) return of(undefined);
-                      partyAccessRoles.forEach( value => {
-                         value.partyId = userRes.partyId;
-                      });
-                      return this.partyAccessRoleRepository.addPartyAccessRole(partyAccessRoles)
-                          .pipe(map( partyAccessRolesRes => {
-                             if (!partyAccessRolesRes) return undefined;
-                             return userRes;
-                          }));
-                  }));
-          }));
+  saveUser (user: User, credential: Credential, partyAccessRoles: PartyAccessRole[], sessionId: string): Observable<User> {
+      if (!credential) {
+          return this.sessionRepository.getSessionById(sessionId)
+              .pipe(switchMap(session => {
+                  user.partyId = session.partyId;
+                  return this.userRepository.saveUser(user);
+              }));
+      } else {
+          credential.password = generate({
+              length: 10,
+              numbers: true
+          });
+          return this.credentialRepository.addCredential(credential)
+              .pipe(switchMap(credentialRes => {
+                  if (!credentialRes) return of(undefined);
+                  user.partyId = credentialRes.credential.partyId;
+                  return this.userRepository.saveUser(user)
+                      .pipe(switchMap(userRes => {
+                          if (!userRes) return of(undefined);
+                          if (!partyAccessRoles || partyAccessRoles.length < 1) return of(userRes);
+                          partyAccessRoles.forEach(value => {
+                              value.partyId = userRes.partyId;
+                          });
+                          return this.partyAccessRoleRepository.addPartyAccessRole(partyAccessRoles)
+                              .pipe(map(partyAccessRolesRes => {
+                                  if (!partyAccessRolesRes) return undefined;
+                                  return userRes;
+                              }));
+                      }));
+              }));
+      }
   }
 
   deleteUser (partyId: string): Observable<number> {
@@ -113,32 +139,35 @@ export class UserOrchestrator {
          }));
     }
 
-  updateUser (partyId: string, user: User, partyAccessRoles: PartyAccessRole[]): Observable<number> {
+  updateUser (partyId: string, user: User, credential: Credential, partyAccessRoles: PartyAccessRole[]): Observable<number> {
       user.username = undefined;
        return this.userRepository.updateUser(partyId, user)
          .pipe(switchMap(numUpdated => {
            if (!numUpdated) return of(undefined);
-           return this.partyAccessRoleRepository.deletePartyAccessRole(partyId)
-              .pipe(switchMap(numRemoved => {
-                  if (!numRemoved) return of(undefined);
-                  return this.partyAccessRoleRepository.addPartyAccessRole(partyAccessRoles)
-                      .pipe(map(next => {
-                          if (!next) return undefined;
-                          return numUpdated;
-                      }));
-             }));
+           return this.credentialRepository.updateUserCredential(partyId, credential)
+               .pipe( switchMap( numUpdated2 => {
+                   if (!numUpdated2) return of(undefined);
+                   if (!partyAccessRoles || partyAccessRoles.length < 1) return of(1);
+                       return this.partyAccessRoleRepository.deletePartyAccessRole(partyId)
+                           .pipe(switchMap(numRemoved => {
+                               return this.partyAccessRoleRepository.addPartyAccessRole(partyAccessRoles)
+                                   .pipe(map(next => {
+                                       return 1;
+                                   }));
+                           }));
+               }));
          }));
     }
 
-  updateUserMe (partyId: string, user: User, credential: Credential): Observable<number> {
-    return this.userRepository.updateUser(partyId, user)
-      .pipe(switchMap(numUpdated => {
-        // if (numUpdated) {
-        //   return this.credentialRepository.updateCredential(partyId, credential);
-        // }else {
-          return of(0);
-        // }
-      }));
-  }
+  // updateUserMe (partyId: string, user: User, credential: Credential): Observable<number> {
+  //   return this.userRepository.updateUser(partyId, user)
+  //     .pipe(switchMap(numUpdated => {
+  //       // if (numUpdated) {
+  //       //   return this.credentialRepository.updateCredential(partyId, credential);
+  //       // }else {
+  //         return of(0);
+  //       // }
+  //     }));
+  // }
 
 }
